Added second half of application questions.
