+++ src/App.js
@@ -14,6 +14,8 @@ import NotFound from "./views/NotFound";
 import NavigationBar from "./components/common/NavigationBar";
 import Footer from "./components/common/Footer";
 import Application from "./views/Application";
+import Info from "./components/application/Info"
+
 
 // Redirect to login page, if use tries to access a restricted page
 function RequireAuth({ children }) {
@@ -58,6 +60,16 @@ export default function App() {
             </RequireAuth>
           }
         />
+
+        <Route
+          path="/info"
+          element={
+            <RequireAuth>
+              <Info />
+            </RequireAuth>
+          }
+        />
+
         <Route path="/login" element={<Login />} />
         <Route path='*' exact={true} element={<NotFound />} />
       </Routes>
